Add upsample layer
